Refactor unit_info.rs from debug/mod.rs
